requirements.txt
streamlit
python-docx
fpdf
PyPDF2
spacy
google-generativeai
python-dotenv